reformated code to be more consistent
