+++ package.json
@@ -14,6 +14,7 @@
     "@emotion/react": "^11.10.6",
     "@types/express": "^4.17.17",
     "@types/jest": "^29.5.0",
+    "@types/morgan": "^1.9.4",
     "@types/node": "^18.15.3",
     "@types/react": "^18.0.28",
     "@types/react-dom": "^18.0.11",
@@ -28,6 +29,7 @@
     "joi": "^17.8.4",
     "mini-css-extract-plugin": "^2.7.5",
     "moment": "^2.29.4",
+    "mongoose": "^7.0.2",
     "morgan": "~1.9.1",
     "prop-types": "^15.8.1",
     "react": "^18.2.0",
@@ -39,6 +41,7 @@
     "webpack-cli": "^5.0.1"
   },
   "devDependencies": {
+    "@types/cookie-parser": "^1.4.3",
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "eslint-config-standard-with-typescript": "^34.0.1",
     "eslint-plugin-import": "^2.25.2",
