+++ src/server/routes/app.ts
@@ -7,4 +7,4 @@ router.get("/", (req, res, next) => {
   res.sendFile("index.html", { root: "public" })
 })
 
-module.exports = router
+export default router
