+++ src/server/routes/journey.ts
@@ -1,0 +1,6 @@
+//Using import syntax instead of require syntax to get typings
+import express from "express";
+
+const router = express.Router()
+
+export default router
